\name{medical}
\alias{medical}
\docType{data}

\title{
medical dataset
}

\description{
extract of modified medical data obtained from physical examination and questionnaires in a human cohort study
}

\usage{data("medical")}

\format{
  A data frame with 40 observations on the following 18 variables.
  \describe{
    \item{\code{medic1}}{a numeric vector}
    \item{\code{medic2}}{a numeric vector}
    \item{\code{medic3}}{a numeric vector}
    \item{\code{medic4}}{a numeric vector}
    \item{\code{medic5}}{a numeric vector}
    \item{\code{medic6}}{a numeric vector}
    \item{\code{medic7}}{a numeric vector}
    \item{\code{medic8}}{a numeric vector}
    \item{\code{medic9}}{a numeric vector}
    \item{\code{medic10}}{a numeric vector}
    \item{\code{medic11}}{a numeric vector}
    \item{\code{medic12}}{a numeric vector}
    \item{\code{medic13}}{a numeric vector}
    \item{\code{medic14}}{a numeric vector}
    \item{\code{medic15}}{a numeric vector}
    \item{\code{medic16}}{a numeric vector}
    \item{\code{medic17}}{a numeric vector}
    \item{\code{medic18}}{a numeric vector}
  }
}

\details{
no details are needed
}

\source{
non-real data
}

\examples{
data(medical)
}

\keyword{datasets}
