\name{nutrition}
\alias{nutrition}
\docType{data}

\title{
nutritional dataset
}

\description{
extract of modified nutritional data obtained by analysis of food questionnaires in a human cohort study
}

\usage{data("nutrition")}
\format{
  A data frame with 40 observations on the following 33 variables.
  \describe{
    \item{\code{nutri1}}{a numeric vector}
    \item{\code{nutri2}}{a numeric vector}
    \item{\code{nutri3}}{a numeric vector}
    \item{\code{nutri4}}{a numeric vector}
    \item{\code{nutri5}}{a numeric vector}
    \item{\code{nutri6}}{a numeric vector}
    \item{\code{nutri7}}{a numeric vector}
    \item{\code{nutri8}}{a numeric vector}
    \item{\code{nutri9}}{a numeric vector}
    \item{\code{nutri10}}{a numeric vector}
    \item{\code{nutri11}}{a numeric vector}
    \item{\code{nutri12}}{a numeric vector}
    \item{\code{nutri13}}{a numeric vector}
    \item{\code{nutri14}}{a numeric vector}
    \item{\code{nutri15}}{a numeric vector}
    \item{\code{nutri16}}{a numeric vector}
    \item{\code{nutri17}}{a numeric vector}
    \item{\code{nutri18}}{a numeric vector}
    \item{\code{nutri19}}{a numeric vector}
    \item{\code{nutri20}}{a numeric vector}
    \item{\code{nutri21}}{a numeric vector}
    \item{\code{nutri22}}{a numeric vector}
    \item{\code{nutri23}}{a numeric vector}
    \item{\code{nutri24}}{a numeric vector}
    \item{\code{nutri25}}{a numeric vector}
    \item{\code{nutri26}}{a numeric vector}
    \item{\code{nutri27}}{a numeric vector}
    \item{\code{nutri28}}{a numeric vector}
    \item{\code{nutri29}}{a numeric vector}
    \item{\code{nutri30}}{a numeric vector}
    \item{\code{nutri31}}{a numeric vector}
    \item{\code{nutri32}}{a numeric vector}
    \item{\code{nutri33}}{a numeric vector}
  }
}

\details{
no details are needed
}

\source{
non-real data
}

\examples{
data(nutrition)
}

\keyword{datasets}
