\name{omics}
\alias{omics}
\docType{data}

\title{
metabolomic dataset
}

\description{
extract of modified metabolomic data obtained by LC-MS analysis of human plasma samples in a cohort study
}

\usage{data("omics")}

\format{
  A data frame with 40 observations on the following 46 variables.
  \describe{
    \item{\code{omic1}}{a numeric vector of relative intensities}
	\item{\code{omic2}}{a numeric vector of relative intensities}
	\item{\code{omic3}}{a numeric vector of relative intensities}
	\item{\code{omic4}}{a numeric vector of relative intensities}
	\item{\code{omic5}}{a numeric vector of relative intensities}
	\item{\code{omic6}}{a numeric vector of relative intensities}
	\item{\code{omic7}}{a numeric vector of relative intensities}
	\item{\code{omic8}}{a numeric vector of relative intensities}
	\item{\code{omic9}}{a numeric vector of relative intensities}
	\item{\code{omic10}}{a numeric vector of relative intensities}
	\item{\code{omic11}}{a numeric vector of relative intensities}
	\item{\code{omic12}}{a numeric vector of relative intensities}
	\item{\code{omic13}}{a numeric vector of relative intensities}
	\item{\code{omic14}}{a numeric vector of relative intensities}
	\item{\code{omic15}}{a numeric vector of relative intensities}
	\item{\code{omic16}}{a numeric vector of relative intensities}
	\item{\code{omic17}}{a numeric vector of relative intensities}
	\item{\code{omic18}}{a numeric vector of relative intensities}
	\item{\code{omic19}}{a numeric vector of relative intensities}
	\item{\code{omic20}}{a numeric vector of relative intensities}
	\item{\code{omic21}}{a numeric vector of relative intensities}
	\item{\code{omic22}}{a numeric vector of relative intensities}
	\item{\code{omic23}}{a numeric vector of relative intensities}
	\item{\code{omic24}}{a numeric vector of relative intensities}
	\item{\code{omic25}}{a numeric vector of relative intensities}
	\item{\code{omic26}}{a numeric vector of relative intensities}
	\item{\code{omic27}}{a numeric vector of relative intensities}
	\item{\code{omic28}}{a numeric vector of relative intensities}
	\item{\code{omic29}}{a numeric vector of relative intensities}
	\item{\code{omic30}}{a numeric vector of relative intensities}
	\item{\code{omic31}}{a numeric vector of relative intensities}
	\item{\code{omic32}}{a numeric vector of relative intensities}
	\item{\code{omic33}}{a numeric vector of relative intensities}
	\item{\code{omic34}}{a numeric vector of relative intensities}
	\item{\code{omic35}}{a numeric vector of relative intensities}
	\item{\code{omic36}}{a numeric vector of relative intensities}
	\item{\code{omic37}}{a numeric vector of relative intensities}
	\item{\code{omic38}}{a numeric vector of relative intensities}
	\item{\code{omic39}}{a numeric vector of relative intensities}
	\item{\code{omic40}}{a numeric vector of relative intensities}
	\item{\code{omic41}}{a numeric vector of relative intensities}
	\item{\code{omic42}}{a numeric vector of relative intensities}
	\item{\code{omic43}}{a numeric vector of relative intensities}
	\item{\code{omic44}}{a numeric vector of relative intensities}
	\item{\code{omic45}}{a numeric vector of relative intensities}
	\item{\code{omic46}}{a numeric vector of relative intensities}
  }
}

\details{
no details are needed
}

\source{
non-real data
}

\examples{
data(omics)
}

\keyword{datasets}
