\name{status}
\alias{status}
\docType{data}

\title{physiopathological status data}

\description{physiopathological status of men in a human cohort study}

\usage{data("status")}

\format{
  A data frame with 40 observations on the following variable.
  \describe{
    \item{\code{status}}{a factor with levels \code{cas} \code{temoin}}
  }
}

\details{no details are needed}

\source{extract of data not yet published}

\examples{
data(status)
}

\keyword{datasets}
